refactor(storage): clean up code formatting and remove unnecessary comments in storage operations
