+++ frontend/src/storage/stores/chord-sheets/operations/delete-all.ts
@@ -1,15 +1,9 @@
-/**
- * Removes all chord sheets from storage
- */
-
-import executeWriteTransaction from '../utils/transactions/write-transaction';
+import { executeWriteTransaction } from "../utils/transactions";
 
 /**
  * @returns Promise that resolves when all data is cleared
  * @throws {DatabaseOperationError} When storage operation fails
  */
 export default async function deleteAllChordSheets(): Promise<void> {
-  return executeWriteTransaction((store) => 
-    store.clear()
-  );
+  return executeWriteTransaction((store) => store.clear());
 }
