+++ frontend/src/storage/stores/chord-sheets/operations/delete-chord-sheet.ts
@@ -1,19 +1,17 @@
-/**
- * Removes a chord sheet from storage
- */
-
-import type { StoredChordSheet } from '../../../types/chord-sheet';
-import executeWriteTransaction from '../utils/transactions/write-transaction';
+import type { StoredChordSheet } from "../../../types/chord-sheet";
+import { executeWriteTransaction } from "../utils/transactions";
 
 /**
  * @param path - Song identifier to remove
  * @throws {DatabaseOperationError} When storage operation fails
  */
-export default async function deleteChordSheet(path: StoredChordSheet['path']): Promise<void> {
+export default async function deleteChordSheet(
+  path: StoredChordSheet["path"]
+): Promise<void> {
   if (!path) {
-    throw new Error('Path is required for delete operation');
+    throw new Error("Path is required for delete operation");
   }
-  
+
   return executeWriteTransaction((store) => {
     return store.delete(path);
   });
