+++ frontend/src/storage/stores/chord-sheets/utils/transactions/index.ts
@@ -1,6 +1,2 @@
-/**
- * Transaction utilities barrel export
- */
-
-export { default as executeReadTransaction } from './read-transaction';
-export { default as executeWriteTransaction } from './write-transaction';
+export { default as executeReadTransaction } from "./read-transaction";
+export { default as executeWriteTransaction } from "./write-transaction";
